Get caches from Director
This commit allows users to set an env var `OSDF_DIRECTOR_URL` that
points `stashcp` to an OSDF Director instance. When this env var is
set, `stashcp` will get its ordered list of caches from the director
instead of directly from the namespaces JSON. This is one of the
first steps in utilizing the new Director infrastructure.