Add a few small changes to make linter happy
The linter was complaining about a variable declaration/assignment spanning two lines,
so I moved them to a single line. It also correctly informed me that I had an issue
with the way I was scoping an err variable that resulted in the new assignment not
being used in an out-of-scope error check.