Implementing and Testing Server-Side HTTP Trailer Mechanism with Staus Parsing
This commit implements the server-side HTTP trailer mechanism, introduces a corresponding test function, and adds a utility to parse the transfer status from the server's response.

In the server-side implementation, we allow a user to set specific headers (X-Transfer-Status and TE) in the request, and then employ chunked encoding in the response. The final chunk returned indicates if an error has occurred.

In order to match this update, we now set these headers in the HTTPRequest. After the response ends, we check for the presence of X-Transfer-Status and the status code. A utility function parseTransferStatus() has been added to extract the status code and status text from the 'X-Transfer-Status' trailer. This function splits the status string into the code and message. If the status code is anything other than 200, an error message is dispatched to the user.

We've also added a new test function 'TestTrailerError' that simulates an HTTP server returning an error trailer, writes some data, and then invokes the 'DownloadHTTP' function. It then checks if the correct error is returned, as expected.

With these changes, we hope to enhance error reporting when there is an IO error in the middle of the response.