Incorporate credential generation object into token acquisition.
This is the first step in using the credential generation data in the
token request -- using the `base_path` from topology to determine what
token scopes we would request.

To make the data structure available in the `oauth2` sub-package (and
avoid a dependency cycle), I had to refactor out the various namespaces
data structures into a standalone sub-package.