Update director to follow cred gen advice
Only recently was the OSDF client re-synced with some of the credential generation advice
supplied by the Topology service. This primarily affected cases where an origin's base
path did not match its local export path (such as in the case of `/chtc/staging/foo`
actually exporting the local path `/staging/foo`). This commit connects the Director into
the reconfigurations so that it will also work in such cases. A parallel commit was also
pushed/merged for the Director to supply the information needed by the OSDF client.